SMTChecker: Ignore errors that might occur during constant evaluation
Previously, if a fatal error would occur during constant evaluation in
the model checker, it would throw an exception which the model checker
would let escape and crash the compiler.
We now use the new API that swallows any errors and the model checker
can continue, simply treating the expression as not constant.